[solvers.hh] Always convert the reduction to double explicitly.

See merge request core/dune-istl!138

+++ dune/istl/solvers.hh
@@ -127,7 +127,7 @@ namespace Dune {
 
       // fill statistics
       res.iterations = i;
-      res.reduction = max_value(def/def0);
+      res.reduction = static_cast<double>(max_value(def/def0));
       res.conv_rate  = pow(res.reduction,1.0/i);
       res.elapsed = watch.elapsed();
 
@@ -1512,7 +1512,7 @@ namespace Dune {
 
       // fill statistics
       res.iterations = i;
-      res.reduction = max_value(def/def0);
+      res.reduction = static_cast<double>(max_value(def/def0));
       res.conv_rate  = pow(res.reduction,1.0/i);
       res.elapsed = watch.elapsed();
